+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 🛠️ Bug fixes:
 
 * Fixed `plotpy.tools.AnnotatedEllipseTool`: `AttributeError` when finalizing the shape
+* `plotpy.widgets.select_with_shape_tool`: added missing `toolbar` and `options` parameters
 
 📚 Documentation: